started redis replacement @ChristoffSchaub NOT COMPILABLE!!!!!!! (lunch time xD)

+++ docs/DatabaseCreate.sql
@@ -50,7 +50,7 @@ CREATE TABLE public.task
     force boolean NOT NULL,
     index_number bigint,
     meta_data jsonb,
-    total_priority integer,
+    total_priority bigint,
     CONSTRAINT task_pkey PRIMARY KEY (id),
     CONSTRAINT group_fk FOREIGN KEY (group_id)
         REFERENCES public."group" (id) MATCH SIMPLE
